fix: 发布帖子的同时删除图片

+++ src/store/edit.js
@@ -63,7 +63,7 @@ const actions = {
     commit(DELETE_EDIT_IMAGES, index)
   },
   async [EDIT_LEAVE] ({ commit, state }) {
-    if (state.imageId.length) {
+    if (state.imageId.length && !isProgressing) {
       await ImageService.delete(state.imageId)
     }
     commit(END_PORGRESSING)
